Problem: delivering extensions to Postgres
Postgres ecosystem has a few hundred extensions and a few
extension/package distribution/installation methods.

[PGXN](http://pgxn.org/), for example, focuses on distribution and meta-information.

[Trunk](https://pgt.dev/) is also a registry and it builds extensions for
a particular platform (Postgres version / operating system
distribution).

(Upcoming) PGA focuses on producing container layers.

[DbDev](https://database.dev/) focuses on TLE (Trusted Language
Extensions)

[pgxman](https://pgxman.com/) is interesting, but still misses a lot of
features, and its open-sourceness is not clear.

Solution: restart the postgres.pm (pgpm) effort.

I started it quite some time ago, experimented with a bunch of ideas,
but it didn't get to fruition.

The main stated goals are now:

* No maintenance on non-breaking extension releases, automatic updates
* Rich package definition language
* Focus on building installable packages (starting with RHEL-compatible
  ones) to facilitate proper deployment

This new version is built as a Ruby gem – with a binary tool – primarily
because:

* It makes a great DSL
* It has enough libraries in the ecosystem to facilitate everythign we
  need

Ruby isn't super fast, but that is not a huge concern for our case.

There are still a lot of rough edges as this is an early prototype, but
it was built with the goal to experiment on a few packages to get to
something _sort of_ working.

A big part of this effort stems from this
[discussion](https://github.com/orgs/omnigres/discussions/648) in the
Omnigres project, but it is very important to make it as usable as
possible for the rest of the ecosystem – which requires a great deal of
flexibility.