Problem: re-initializing mock roots
Every time we have a new container with pgpm, it has to reinitialize its
roots for building RPMs using mock. It takes a good while.

Solution: pre-initialize roots during image preparation

Note: this requires image building to be done in a privileged mode. It's
been documented how to do this in the Dockerfile.