Problem: failing to package non-native-code Omnigres extensions
It is failing to build `package_NAME_extension`

Solution: only build present targets and exclude `omni` if not requested

Extension without compilable sources do not have this target but we also
need to build dependencies.

As a reminder, we try to avoid building `omni` because it is
automatically injected into every build whether we need it or not and
older versions of it might not compile even though it is not necessary
for the package we're trying build – preventing it from being published.