Problem: can't build omnigres/omni_schema
It is detected to be broken. However, I don't believe that's true!

Solution: ensure we use the latest build system

The reason why it wasn't doing this before is that it saw changes
between the versions. However, the changes only adjusted the changelog,
which is not really modifying the actual extension.