Problem: some Omnigres extensions still won't get to building
Particularly, it seems to be the case for omni_schema.

Solution: don't use merges that were not directly merged into master

When Omnigres merges are worked on, some other branches may sometimes
get merged into them explictly. We were detecting those as actual merges
when it comes to detecting published versions.

However, consider this scenario: we bump up a version in a merge that's
merged into a PR but then change the version. We don't want that to be
detected as a legitimate version.

Also, new version mapping is much faster. A little cleaner, too.