Problem: wrong function names in .sql files
Sometimes we get really weird function names in resulting .sql files,
like here:

```
ERROR:  could not find function "halfvec_subvector--0.8.0.so" in file "/usr/pgsql-17/lib/vector--0.8.0.so"
```

Solution: ensure we actually search&replace file names
and not function names.

We do so by requiring a preceding slash. Hopefully this is good enough
for now. In the future, maybe we can switch to a Ruby script that'll
parse SQL files?